check that envfile exists

+++ resources/systemd/update_tms_image_symlink.sh
@@ -2,7 +2,8 @@
 set -euo pipefail
 
 # ------------------------------------------------------------------------------
-# Script Name: update_tms_image_symlink.sh
+# update_tms_image_symlink.sh
+#
 # Description: Updates the Apptainer container symlink to a specified image
 #              version stored on CVMFS.
 #
@@ -11,16 +12,10 @@ set -euo pipefail
 # Arguments:
 #   TMS_IMAGE_TAG - The version tag of the TMS image to be used.
 #
-# Behavior:
-#   - Ensures the script is run from either 'tms/' or 'tms-dev/' directory.
-#   - Validates that a TMS image tag is provided as an argument.
-#   - Checks if the specified image exists on CVMFS.
-#   - Updates the symlink to point to the specified image.
-#   - Touches `envfile` to trigger a systemd service restart.
-#
 # Exit Codes:
 #   1 - Invalid working directory or missing argument.
 #   2 - Specified image version not found on CVMFS.
+#   3 - `envfile` is missing, preventing symlink update.
 #
 # Example:
 #   ./update_tms_image_symlink.sh 0.1.52
@@ -44,6 +39,7 @@ tms_image_tag="$1"
 
 ################################################################################
 # constants
+
 readonly cvmfs_base="/cvmfs/icecube.opensciencegrid.org/containers/ewms/observation-management-service/ewms-task-management-service"
 
 readonly full_image_path="$cvmfs_base:$tms_image_tag"
@@ -52,12 +48,18 @@ if [[ ! -d "$full_image_path" ]]; then
     exit 2
 fi
 
+readonly envfile="./envfile"
+if [[ ! -f $envfile ]]; then
+    echo "Error: './envfile' is missing. Not updating symlink."
+    exit 3
+fi
+
 ################################################################################
 # update!
 
 ln -snf "$full_image_path" "./apptainer_container_symlink"
 
 # Touch envfile so systemd restarts the app
-touch "./envfile"
+touch $envfile
 
 echo "Successfully updated symlink to: $(readlink -f ./apptainer_container_symlink)"
